refactor: address PR #103 review — split CLI + backtrace, remove dead code
Split monolithic cli/__init__.py (979 lines) into parser.py + dispatch.py
(<400 lines each). Extract backtrace regex patterns into backtrace_patterns.py.
Remove dead probe-rs/SWO CLI dispatch, unused imports, dead except clauses.
Add magic number comments in jlink_rtt.py and WHY comments on regex patterns.
