+++ eab/reset_reason.py
@@ -13,7 +13,7 @@ Maintains reset history in status.json and alerts on unexpected resets.
 from __future__ import annotations
 
 import re
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from datetime import datetime
 from typing import Optional
